fix max line lint error for vision
