+++ libs/interpret-vision/src/lib/VisionExplanationDashboard/Controls/DataCharacteristics.tsx
@@ -15,45 +15,18 @@ import { localization } from "@responsible-ai/localization";
 import React from "react";
 
 import { dataCharacteristicsStyles } from "./DataCharacteristics.styles";
+import {
+  defaultState,
+  getLabelVisibility,
+  IDataCharacteristicsProps,
+  IDataCharacteristicsState,
+  labelTypes,
+  processData,
+  stackTokens
+} from "./DataCharacteristicsHelper";
 import { DataCharacteristicsLegend } from "./DataCharacteristicsLegend";
 import { DataCharacteristicsRow } from "./DataCharacteristicsRow";
 
-export interface IDataCharacteristicsProps {
-  data: IVisionListItem[];
-  imageDim: number;
-  numRows: number;
-  selectItem(item: IVisionListItem): void;
-}
-export interface IDataCharacteristicsState {
-  columnCount: number[];
-  dropdownOptionsPredicted: IDropdownOption[];
-  dropdownOptionsTrue: IDropdownOption[];
-  itemsPredicted: Map<string, IVisionListItem[]>;
-  itemsTrue: Map<string, IVisionListItem[]>;
-  labelType: string;
-  labelTypeDropdownOptions: IDropdownOption[];
-  labelVisibilitiesPredicted: Map<string, boolean>;
-  labelVisibilitiesTrue: Map<string, boolean>;
-  renderStartIndex: number[];
-  selectedKeysPredicted: string[];
-  selectedKeysTrue: string[];
-  showBackArrow: boolean[];
-}
-
-interface IItemsData {
-  dropdownOptions: IDropdownOption[];
-  items: Map<string, IVisionListItem[]>;
-  labelVisibilities: Map<string, boolean>;
-  selectedKeys: string[];
-}
-
-const stackTokens = { childrenGap: "l1" };
-const labelTypes = {
-  predictedY: "predictedY",
-  trueY: "trueY"
-};
-
-const SelectAllKey = "selectAll";
 export class DataCharacteristics extends React.Component<
   IDataCharacteristicsProps,
   IDataCharacteristicsState
@@ -62,21 +35,7 @@ export class DataCharacteristics extends React.Component<
   public constructor(props: IDataCharacteristicsProps) {
     super(props);
     this.rowHeight = 0;
-    this.state = {
-      columnCount: [],
-      dropdownOptionsPredicted: [],
-      dropdownOptionsTrue: [],
-      itemsPredicted: new Map(),
-      itemsTrue: new Map(),
-      labelType: labelTypes.predictedY,
-      labelTypeDropdownOptions: [],
-      labelVisibilitiesPredicted: new Map(),
-      labelVisibilitiesTrue: new Map(),
-      renderStartIndex: [],
-      selectedKeysPredicted: [],
-      selectedKeysTrue: [],
-      showBackArrow: []
-    };
+    this.state = defaultState;
   }
 
   public componentDidMount(): void {
@@ -84,7 +43,9 @@ export class DataCharacteristics extends React.Component<
       { key: labelTypes.predictedY, text: labelTypes.predictedY },
       { key: labelTypes.trueY, text: labelTypes.trueY }
     ];
-    this.setState({ labelTypeDropdownOptions });
+    this.setState({
+      labelTypeDropdownOptions
+    });
     this.processData();
   }
 
@@ -205,6 +166,10 @@ export class DataCharacteristics extends React.Component<
     );
   }
 
+  private processData = (): void => {
+    this.setState(processData(this.props.data));
+  };
+
   private onRenderCell = (
     item?: IVisionListItem | undefined
   ): React.ReactElement => {
@@ -236,102 +201,6 @@ export class DataCharacteristics extends React.Component<
     );
   };
 
-  private generateItems = (
-    type: string,
-    examples: IVisionListItem[]
-  ): IItemsData => {
-    const dropdownOptions: IDropdownOption[] = [];
-    const items: Map<string, IVisionListItem[]> = new Map();
-    const labelVisibilities: Map<string, boolean> = new Map();
-    const selectedKeys: string[] = [];
-    examples.forEach((example) => {
-      const label = example[type].toString();
-      if (!label || !items) {
-        return;
-      }
-      if (!selectedKeys.includes(label)) {
-        dropdownOptions.push({ key: label, text: label });
-        selectedKeys.push(label);
-        labelVisibilities.set(label, true);
-      }
-      if (items.has(label)) {
-        const arr = items.get(label);
-        if (!arr) {
-          return;
-        }
-        arr.push(example);
-        items.set(label, arr);
-      } else {
-        const arr: IVisionListItem[] = [];
-        arr.push(example);
-        items.set(label, arr);
-      }
-    });
-    selectedKeys.push(SelectAllKey);
-    dropdownOptions.push({
-      key: SelectAllKey,
-      text: localization.InterpretVision.Dashboard.selectAll
-    });
-    dropdownOptions.sort((a, b) => {
-      if (b.key === SelectAllKey) {
-        return 1;
-      }
-      if (a.key === SelectAllKey || a.key < b.key) {
-        return -1;
-      }
-      return 1;
-    });
-    return {
-      dropdownOptions,
-      items,
-      labelVisibilities,
-      selectedKeys
-    };
-  };
-
-  private processData = (): void => {
-    const examples: IVisionListItem[] = this.props.data;
-    const columnCount: number[] = [];
-    const renderStartIndex: number[] = [];
-    const showBackArrow: boolean[] = [];
-
-    const predicted: IItemsData = this.generateItems(
-      labelTypes.predictedY,
-      examples
-    );
-    const dropdownOptionsPredicted: IDropdownOption[] =
-      predicted.dropdownOptions;
-    const itemsPredicted: Map<string, IVisionListItem[]> = predicted.items;
-    const labelVisibilitiesPredicted: Map<string, boolean> =
-      predicted.labelVisibilities;
-    const selectedKeysPredicted: string[] = predicted.selectedKeys;
-
-    const trues: IItemsData = this.generateItems(labelTypes.trueY, examples);
-    const dropdownOptionsTrue: IDropdownOption[] = trues.dropdownOptions;
-    const itemsTrue: Map<string, IVisionListItem[]> = trues.items;
-    const labelVisibilitiesTrue: Map<string, boolean> = trues.labelVisibilities;
-    const selectedKeysTrue: string[] = trues.selectedKeys;
-
-    selectedKeysPredicted.forEach(() => {
-      renderStartIndex.push(0);
-      showBackArrow.push(false);
-      columnCount.push(0);
-    });
-    this.setState({
-      columnCount,
-      dropdownOptionsPredicted,
-      dropdownOptionsTrue,
-      itemsPredicted,
-      itemsTrue,
-      labelVisibilitiesPredicted,
-      labelVisibilitiesTrue,
-      renderStartIndex,
-      selectedKeysPredicted,
-      selectedKeysTrue,
-      showBackArrow
-    });
-  };
-
   private onLabelTypeChange = (
     _event: React.FormEvent<HTMLDivElement>,
     item?: IDropdownOption<any> | undefined
@@ -345,81 +214,10 @@ export class DataCharacteristics extends React.Component<
     _event: React.FormEvent<HTMLDivElement>,
     item?: IDropdownOption<any> | undefined
   ): void => {
-    if (item) {
-      const predicted = this.state.labelType === labelTypes.predictedY;
-
-      if (item.key === SelectAllKey) {
-        const dropdownOptions = predicted
-          ? this.state.dropdownOptionsPredicted
-          : this.state.dropdownOptionsTrue;
-        const selectedKeys: string[] = [];
-        const labelVisibilities: Map<string, boolean> = new Map();
-        if (
-          (predicted &&
-            this.state.selectedKeysPredicted.length >=
-              dropdownOptions.length - 1) ||
-          (!predicted &&
-            this.state.selectedKeysTrue.length >= dropdownOptions.length - 1)
-        ) {
-          dropdownOptions.forEach((option, index) => {
-            if (index !== 0) {
-              labelVisibilities.set(option.key.toString(), false);
-            }
-          });
-        } else {
-          dropdownOptions.forEach((option, index) => {
-            if (index !== 0) {
-              selectedKeys.push(option.key.toString());
-              labelVisibilities.set(option.key.toString(), true);
-            }
-          });
-          selectedKeys.push(SelectAllKey);
-        }
-        if (predicted) {
-          this.setState({
-            labelVisibilitiesPredicted: labelVisibilities,
-            selectedKeysPredicted: [...selectedKeys]
-          });
-        } else {
-          this.setState({
-            labelVisibilitiesTrue: labelVisibilities,
-            selectedKeysTrue: [...selectedKeys]
-          });
-        }
-        return;
-      }
-
-      let selectedKeys: string[] = [];
-      let labelVisibilities: Map<string, boolean> = new Map();
-
-      if (predicted) {
-        selectedKeys = this.state.selectedKeysPredicted;
-        labelVisibilities = this.state.labelVisibilitiesPredicted;
-      } else {
-        selectedKeys = this.state.selectedKeysTrue;
-        labelVisibilities = this.state.labelVisibilitiesTrue;
-      }
-
-      if (selectedKeys.includes(SelectAllKey)) {
-        selectedKeys.splice(selectedKeys.indexOf(SelectAllKey), 1);
-      }
-      const key = item.key.toString();
-      selectedKeys.includes(key)
-        ? selectedKeys.splice(selectedKeys.indexOf(key), 1)
-        : selectedKeys.push(key);
-      labelVisibilities.set(key, !labelVisibilities.get(key));
-      if (predicted) {
-        this.setState({
-          labelVisibilitiesPredicted: labelVisibilities,
-          selectedKeysPredicted: [...selectedKeys]
-        });
-      } else {
-        this.setState({
-          labelVisibilitiesTrue: labelVisibilities,
-          selectedKeysTrue: [...selectedKeys]
-        });
-      }
+    if (!item) {
+      return;
     }
+    this.setState(getLabelVisibility(item, this.state));
   };
 
   private sortKeys(keys: string[]): string[] {
